LayoutTests/fast/canvas/canvas-scale-drawImage-shadow-expected.txt

Ensure correct behavior of canvas with drawImage+shadow after scaling. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 83
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 83
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 53
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 24
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 24
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 24
